Minimize activation function for Zowe Explorer
